pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aumos-audio-engine"
version = "0.1.0"
description = "Synthetic audio generation with speaker de-identification, MNPI detection, and biometric-free voice synthesis for financial sector"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [{ name = "AumOS Enterprise" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aumos-common>=0.1.0",
    "aumos-proto>=0.1.0",
    "fastapi>=0.110.0",
    "pydantic>=2.6.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "TTS>=0.22.0",
    "audiocraft>=1.2.0",
    "faster-whisper>=1.0.0",
    "torchaudio>=2.2.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "testcontainers[postgres,kafka,redis]>=3.7.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/aumos_audio_engine"]

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "ANN", "B", "A", "COM", "C4", "PT", "RUF"]
ignore = ["ANN101", "ANN102", "COM812"]

[tool.ruff.lint.isort]
known-first-party = ["aumos_audio_engine"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
exclude = ["tests/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=aumos_audio_engine --cov-report=term-missing --cov-fail-under=80"
